… src/System.Private.CoreLib/shared/System/Collections/Generic/ArraySortHelper.cs
… src/classlibnative/bcltype/arrayhelpers.cpp
… src/classlibnative/bcltype/arrayhelpers.h