@GSPP 2019-11-06
It is possible that `typeof(T[]) != array.GetType()`. This happens when passing in, for example, `(uint[])(object)new int[10]`. The CLR type system allows treating an `int[]` and `uint[]`.

I wonder if there is a behavior change here if previously the native code used the runtime type of the array and now the C# code uses the type of the generic argument.