@jkotas 2019-11-06
All these cases should work fine. The switch is done on TypeCode. It that prevents `int[]` to be confused with `uint[]`.