@adamsitnik 2019-11-06
@nietras @stephentoub great work!

/cc @damageboy who is working on a vectorized version of Array.Sort ;)